+++ examples/test/sizedint/test_sizedint_bad_byte.pas
@@ -1,0 +1,6 @@
+program TestSizedIntBadByte;
+var
+    b: byte;
+begin
+    b := 300;
+end.