+++ examples/test/sizedint/test_sizedint_bad_shortint.pas
@@ -1,0 +1,6 @@
+program TestSizedIntBadShortint;
+var
+    s: shortint;
+begin
+    s := 200;
+end.