+++ examples/test/sizedint/test_sizedint_bad_word.pas
@@ -1,0 +1,6 @@
+program TestSizedIntBadWord;
+var
+    w: word;
+begin
+    w := 70000;
+end.